Functions and structs

From $mpccmodel_common.jl$

Common structs, etc.

MPCCModelManager.MPCCDimSpec — Type

MPCCDimSpec

Determines the dimensions of a program.

n: number of spatial dimensions

q: columns in F (number of complementarity contraints)

l: row in F (number of expressions in each complementarity contraint)

me: numeber of equality constraints

me: numeber of inequality constraints

r: number of continuous parameters

s: number of integer parameters

MPCCModelManager.MPCCDefinition — Type

MPCCDefinition

Definition of the MPCC/NLP containing Symbolics Num expressions for f, ce, ci, and F.

f: objective function

ce: vector of equality constraints ( == 0 )

ci: vector of inequality constraints ( >= 0 )

F: matrix l-by-q of complementarity constraints ( >= 0 )

MPCCModelManager.MPCCFunctions — Type

MPCCFunctions

A MPCCDefinition compiled into functions. This contains both standard and mutating functions, and individual elements encoded as a function within a vector or matrix.

MPCCModelManager.MPCCParameterisationDefn — Type

MPCCParameterisationDefn

The definition for a parameterisation of a model, i.e. how the pr depends on t.

pr: vector of Num expressions

tspan: closed t interval

descr: description

MPCCModelManager.MPCCParameterisationFunctions — Type

MPCCParameterisationFunctions

Compilation of MPCCParameterisationFunctions into Julia functions, including derivative.

MPCCModelManager.MPCCParameterisations — Type

MPCCParameterisations

Collection of Num t along with definitions of parameterisations and their compilation.

MPCCModelManager.MPCCModelConfig — Type

MPCCModelConfig

Model config contains the Num variables for x, pr, ps along with the MPCCDimSpec, definition, compiled definition, any test vectors, any known solutions, and parameterisations. Basically everything that specifies a model.

Note re parameterisations: if using a static model with r=0, then this should be set to missing

MPCCModelManager.MPCCModelDenseForwardDiff — Type

MPCCModelDenseForwardDiff

A MPCCModelConfig compiled to produce the derivatives we want. This version uses ForwardDiff.jl and produces dense matrices.

MPCCModelManager.MPCCActivesetCnstrType — Type

MPCCActivesetCnstrType

Enumeration of the different possible constraint types: equality, inequality, and each individual constraint in a complementarity.

MPCCModelManager.MPCCConstraintSet — Type

MPCCConstraintSet

A constraint set, e.g. for storing a working or active set. It works by collecting a vector of base indices. By a standard definition, each constraint in a problem has a unique base index.

So we (only) need to store the problem dimension and the vector of BIs. There are functions that encode the logic to do the opposite conversion. In practice, we actually save the model config because for display purposes, it allows us to list the actual constraint expressions.

From $mpccmodel_proc_model.jl$

These functions are used to process a model definition with end point producing the model config struct which includes closures to construct f, ce, ci, F functions, parametrisations, etc.

MPCCModelManager.mpccmodel_build_sym_nums — Function

mpccmodel_build_sym_nums(dimspec::MPCCDimSpec)

Returns Symbolics.jl Num variable that correspond to the provided dimspec. See MPCCDimSpec definition for what each field means.

MPCCModelManager.mpccmodel_build_fn_from_defn — Function

mpccmodel_build_fn_from_defn(dimspec, defn, x, pr, ps)

Assemble f, ce, ci, and F Julia native functions from Symbolic definition and returns both non-indexed and indexed variants. Generally no need to call directly as this is called from mpccmodel_load_defn_from_file().

NOTE: To understand how this works, one should also play with a few examples ofbuild_function() in Symbolics.jl. The output from build_function() is a 2-element vector whereby the first element is the standard function and the second is a mutating version of the same function. Hence the formulation in the code here.

Arguments

dimspec::MPCCDimSpec: The dimension specifications for the proram.

defn::MPCCDefinition: The Symbolics.jl Num definitions of the expressions.

x, pr, ps: Symbolics.jl Num variables for x, pr, ps.

MPCCModelManager.mpccmodel_build_parameterisation — Function

mpccmodel_build_parameterisation(defns, t)

Build parameterisation from defns and Symbolics Num t, returns struct with Julia native functions including derivatives.

Missing docstring.

Missing docstring for MPCCModelManager.mpccmodel_construct_config. Check Documenter's build log for details.

From $mpccmodel_calc_forwarddiff_dense.jl$

This construsts the full gamut of functions as closures using ForwardDiff.jl.

MPCCModelManager.mpccmodel_setup_forwarddiff_dense — Function

mpccmodel_setup_forwarddiff_dense(config::MPCCModelConfig)

Accepts model config and returns a struct with all the nice Julia functions to calculate Jacobians, Hessians, and parametric derivatives using ForwardDiff.jl.

Both standard and mutating functions are constructed. Most of the functions accept an index as the last argument to specify a subset of the constraints. For F constraints, this should be a vector of either tuple pairs or CartesianIndexes.

From $mpccmodel_constraint_sets.jl$

This code is used when storing the active set description. It allow manipulation of an active set description in a storage efficient manner.

MPCCModelManager.cs_cpair_to_bi — Function

cs_cpair_to_bi(dimspec::MPCCDimSpec, ctype::MPCCActivesetCnstrType, el::MPCCIdxElement)

Convert from a 'cpair' of constraint type (CE, CI, F) and element index, to a base index in the defined ordering. Only depends on the problem dimensions, not an actual constraint set.

MPCCModelManager.cs_bi_to_cpair — Function

cs_bi_to_cpair(dimspec::MPCCDimSpec, el::Int64; b_F_cart=true)

Convert a base index, in the defined ordering, to a cpair of constraint type (CE, CI, F) and element index. Only depends on problem dimensions, not on an actual constraint set.

MPCCModelManager.cs_cnstr_el_to_bi — Function

cs_cnstr_el_to_bi(cnstr_idxs::MPCCConstraintSet, el::Int64)

Convert element of a constraint set to the base index. This DOES depend on the constraint set.

MPCCModelManager.cs_bi_to_cnstr_el — Function

cs_bi_to_cnstr_el(cnstr_idxs::MPCCConstraintSet, bi_idx::Int64)

Convert a base index to the constraint set index. This may return missing since the specified BI might not be in the constraint set (not active). This DOES depend on the constraint set.

MPCCModelManager.cs_cnstr_el_to_cpair — Function

cs_cnstr_el_to_cpair(cnstr_idxs::MPCCConstraintSet, el::Int64; b_F_cart=true)

Convert element el of constraint set to the cpair format.

MPCCModelManager.cs_cpair_to_cnstr_el — Function

cs_cpair_to_cnstr_el(cnstr_idxs::MPCCConstraintSet, ctype::MPCCActivesetCnstrType, el::MPCCIdxElement)

Convert cpair element of a constraint set to its element in the constraint set (at which place is it in the bi index)

MPCCModelManager.cs_cnstr_idxs_add_by_bi! — Function

cs_cnstr_idxs_add_by_bi!(cnstr_idxs::MPCCConstraintSet, bi_idx::Int64)

Add element into constraint set by its base index.

MPCCModelManager.cs_cnstr_idxs_add_by_cpair! — Function

cs_cnstr_idxs_add_by_cpair!(cnstr_idxs::MPCCConstraintSet, ctype::MPCCActivesetCnstrType, el::MPCCIdxElement)

Add element into the constraint set by its cpair.

MPCCModelManager.cs_cnstr_idxs_del_by_cnstr_el! — Function

cs_cnstr_idxs_del_by_cnstr_el!(cnstr_idxs::MPCCConstraintSet, el::Int64)

Delete element of constraint set by its element number.

MPCCModelManager.cs_cnstr_idxs_del_by_bi! — Function

cs_cnstr_idxs_del_by_bi!(cnstr_idxs::MPCCConstraintSet, bi_idx::Int64)

Delete element of a constraint set by its base index.

MPCCModelManager.cs_cnstr_idxs_del_by_cpair! — Function

cs_cnstr_idxs_del_by_cpair!(cnstr_idxs::MPCCConstraintSet, ctype::MPCCActivesetCnstrType, el::MPCCIdxElement)

Delete constraint set element by its cpair.

MPCCModelManager.cs_cnstr_get_as_bitmask — Function

cs_cnstr_get_as_bitmask(cnstr_idxs::MPCCConstraintSet)

Get the constraint set expressed as three bitmasks: BitVector for CE and CI; BitMatrix for F. Returned as NamedTuple.

MPCCModelManager.cs_cnstr_build_from_bitmask — Function

cs_cnstr_build_from_bitmask(dimspec::MPCCDimSpec, ce_actv::BitVector, ci_actv::BitVector, F_actv::BitMatrix)

Build whole constraint set from bitmasks of the contraints for each type.

MPCCModelManager.cs_cnstr_build_all_active — Function

cs_cnstr_build_all_active(dimspec::MPCCDimSpec)

Build a constraint set with all constraints included (active).

MPCCModelManager.cs_cnstr_get_Fq_count — Function

cs_cnstr_get_Fq_count(cnstr_idxs::MPCCConstraintSet)

Return the number of active constraints in each column q of the F constraints.

MPCCModelManager.cs_cnstr_get_all_active_bi — Function

cs_cnstr_get_all_active_bi(cnstr_idxs::MPCCConstraintSet)

Returns BI entries for all contraints in cnstr_idxs, separated into type in a named tuple.