Functions and structs

MPCCModelManager.mpccmodel_build_fn_from_defn — Function

mpccmodel_build_fn_from_defn(dimspec, defnnum, x, pr, ps)

Assemble f, ce, ci, and F Julia native functions from Symbolic definition and returns both non-indexed and indexed variants. Generally no need to call directly as this is called from mpccmodel_load_defn_from_file().

NOTE: To understand how this works, one should also play with a few examples ofbuild_function() in Symbolics.jl. The output from build_function() is a 2-element vector whereby the first element is the standard function adn the second is a mutating version of the same function. Hence the formulation in the code here.

Arguments

dimspec::MPCCDimSpec: The dimension specifications for the proram.

defnnum::MPCCDefinition: The Symbolics.jl Num definitions of the expressions.

x, pr, ps: Symbolics.jl Num variables for x, pr, ps.